test_polymarket_connectivity/requirements.txt
py-clob-client
python-dotenv